Move Device and AudioDeviceToRecord classes to their own files
