+++ Casterr.RecorderLib/FFmpeg/Device.cs
@@ -1,0 +1,14 @@
+namespace Casterr.RecorderLib.FFmpeg
+{
+  public class Device
+  {
+    // Source Number
+    public int ID { get; set; }
+
+    // Name of device
+    public string Name { get; set; }
+
+    // Is device an input
+    public bool IsInput { get; set; }
+  }
+}
