+++ Casterr.RecorderLib/FFmpeg/DeviceManager.cs
@@ -6,18 +6,6 @@ using System.Runtime.InteropServices;
 
 namespace Casterr.RecorderLib.FFmpeg
 {
-  public class Device
-  {
-    // Source Number
-    public int ID { get; set; }
-
-    // Name of device
-    public string Name { get; set; }
-
-    // Is device an input
-    public bool IsInput { get; set; }
-  }
-
   public class DeviceManager
   {
     public string DesktopVideoDevice = "screen-capture-recorder";
