+++ Casterr.SettingsLib/AudioDeviceToRecord.cs
@@ -1,0 +1,8 @@
+namespace Casterr.SettingsLib
+{
+  public class AudioDeviceToRecord
+  {
+    public int SourceNumber { get; set; }
+    public string Name { get; set; }
+  }
+}
